+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvGenders.java
@@ -3,6 +3,6 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvGenders(
-    @CsvBindByName(column = "id") int id,
+    @CsvBindByName(column = "id") Long id,
     @CsvBindByName(column = "identifier") String identifier
 ) {}
