+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvLocation.java
@@ -3,7 +3,7 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvLocation(
-		@CsvBindByName(column = "id") int id,
+		@CsvBindByName(column = "id") Long id,
 		@CsvBindByName(column = "region_id") Integer regionId,
 		@CsvBindByName(column = "identifier") String identifier,
 		@CsvBindByName(column = "name") String name,
