+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvPokemonDexNumbers.java
@@ -3,7 +3,7 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvPokemonDexNumbers(
-    @CsvBindByName(column = "species_id") int speciesId,
-    @CsvBindByName(column = "pokedex_id") int pokedexId,
+    @CsvBindByName(column = "species_id") Long speciesId,
+    @CsvBindByName(column = "pokedex_id") Long pokedexId,
     @CsvBindByName(column = "pokedex_number") int pokedexNumber
 ) {}
