+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvPokemonFormGenerations.java
@@ -3,7 +3,7 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvPokemonFormGenerations(
-    @CsvBindByName(column = "pokemon_form_id") int pokemonFormId,
-    @CsvBindByName(column = "generation_id") int generationId,
+    @CsvBindByName(column = "pokemon_form_id") Long pokemonFormId,
+    @CsvBindByName(column = "generation_id") Long generationId,
     @CsvBindByName(column = "game_index") int gameIndex
 ) {}
