+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvType.java
@@ -3,6 +3,6 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvType(
-		@CsvBindByName(column = "id") int pokemonId,
-		@CsvBindByName(column = "identifier") int typeId
+		@CsvBindByName(column = "id") Long pokemonId,
+		@CsvBindByName(column = "identifier") Long typeId
 ) {}
