+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvTypeGameIndices.java
@@ -3,7 +3,7 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvTypeGameIndices(
-    @CsvBindByName(column = "type_id") int typeId,
-    @CsvBindByName(column = "generation_id") int generationId,
+    @CsvBindByName(column = "type_id") Long typeId,
+    @CsvBindByName(column = "generation_id") Long generationId,
     @CsvBindByName(column = "game_index") int gameIndex
 ) {}
