+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvVersions.java
@@ -3,8 +3,8 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvVersions(
-    @CsvBindByName(column = "id") int id,
-    @CsvBindByName(column = "version_group_id") int versionGroupId,
+    @CsvBindByName(column = "id") Long id,
+    @CsvBindByName(column = "version_group_id") Long versionGroupId,
     @CsvBindByName(column = "identifier") String identifier,
     @CsvBindByName(column = "name") String name
 ) {}
